Move legacy internal macros into `update-cfn` dir and give them deploy order precedence - Pass `deployStage` to Inventory
