+++ src/sam/update-cfn/_legacy-api/add-static-proxy.js
