+++ src/sam/update-cfn/_legacy-api/get-api-properties.js
