+++ src/sam/update-cfn/_static/index.js
